requirements.txt
asgiref==3.5.2
boto3==1.26.29
botocore==1.29.29
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
cryptography==38.0.4
Django==4.1.4
django-cors-headers==3.13.0
django-enumchoicefield==3.0.1
djangorestframework==3.14.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
jmespath==1.0.1
Pillow==9.3.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
rfc3986==1.5.0
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
